backend/requirements.txt
asgiref==3.11.1
Django==6.0.3
sqlparse
djangorestframework
djangorestframework-simplejwt
django-cors-headers
psycopg2-binary
django-environ
django-jazzmin
Pillow
celery[redis]
requests
gunicorn
africastalking
reportlab
whitenoise
pygments
sentry-sdk
